requirements.txt
fastapi
uvicorn
google-genai
httpx
python-dotenv
jinja2
python-multipart
